Expand data node constructor for creating and loading from storages.
This moves the functionality from the load() method to be used from
__init__. The implementation is in _init_from_storage(), so subclasses
do not need to override __init__ entirely.
Also, this provides a common, backend-agnostic interface for creating
new data node's storage objects, supporting additional arguments in the
process. This is needed whenever a backend requires parameters for
creating the storage obeject.